Add more entities to domain
