+++ src/domain/aggregated/entities/MappedCategoryOption.ts
@@ -1,0 +1,6 @@
+export interface MappedCategoryOption {
+    dataElement: string;
+    categoryOptions: string[];
+    mappedOptionCombo: string;
+    category: string;
+}
