+++ src/domain/common/entities/NamedRef.ts
@@ -1,0 +1,5 @@
+import { Ref } from "./Schemas";
+
+export interface NamedRef extends Ref {
+    name: string;
+}
