+++ src/domain/common/entities/SharingSetting.ts
@@ -1,0 +1,6 @@
+export interface SharingSetting {
+    access: string;
+    displayName: string;
+    id: string;
+    name: string;
+}
