+++ src/domain/common/entities/Validations.ts
