+++ src/types/synchronization.d.ts
@@ -1,6 +1,7 @@
 import { D2ModelSchemas, Ref } from "../types/d2-api";
 import SyncReport from "../models/syncReport";
 import { DataImportParams, MetadataImportParams } from "./d2";
+import { SynchronizationRule } from "./SynchronizationRule";
 
 //TODO: Review this to move it to domain
 
@@ -39,31 +40,6 @@ export interface ExportBuilder {
     includeSharingSettings: boolean;
 }
 
-export type SynchronizationReportStatus = "READY" | "RUNNING" | "FAILURE" | "DONE";
-
-export interface SynchronizationReport {
-    id: string;
-    date?: Date;
-    user: string;
-    status: SynchronizationReportStatus;
-    types: string[];
-    syncRule?: string;
-    deletedSyncRuleLabel?: string;
-    type: SyncRuleType;
-    dataStats?: AggregatedDataStats[] | EventsDataStats[];
-}
-
-export interface AggregatedDataStats {
-    dataElement: string;
-    count: number;
-}
-
-export interface EventsDataStats {
-    program: string;
-    count: number;
-    orgUnits: string[];
-}
-
 export interface NestedRules {
     [metadataType: string]: string[][];
 }
@@ -72,45 +48,4 @@ export interface SynchronizationState {
     message?: string;
     syncReport?: SyncReport;
     done?: boolean;
-}
-
-interface NamedRef extends Ref {
-    name: string;
-}
-
-export interface SynchronizationRule {
-    id: string;
-    name: string;
-    code?: string;
-    created: Date;
-    description?: string;
-    builder: SynchronizationBuilder;
-    enabled: boolean;
-    lastExecuted?: Date;
-    lastUpdated: Date;
-    lastUpdatedBy: NamedRef;
-    frequency?: string;
-    publicAccess: string;
-    user: NamedRef;
-    userAccesses: SharingSetting[];
-    userGroupAccesses: SharingSetting[];
-    type: SyncRuleType;
-}
-
-export type SynchronizationRuleMain = Omit<SynchronizationRule, DetailsKeys>;
-
-export type SyncRuleType = "metadata" | "aggregated" | "events" | "deleted";
-
-export interface SharingSetting {
-    access: string;
-    displayName: string;
-    id: string;
-    name: string;
-}
-
-export interface CategoryOptionAggregationBuilder {
-    dataElement: string;
-    categoryOptions: string[];
-    mappedOptionCombo: string;
-    category: string;
 }
